refactored a lot
switched to esmodule
splited app.js up more (primarly db)
