+++ app.js
@@ -6,32 +6,45 @@
 
 //* ----------------- IMPORTS -----------------
 
-const { exec } = require("child_process");
-const chokidar = require("chokidar");
-const _colors = require("colors");
-const dotenv = require("dotenv");
-const express = require("express");
-const fs = require("fs");
-const helmet = require("helmet");
-const http = require("http");
-const io = require("@pm2/io");
-const { networkInterfaces } = require("os");
-const path = require("path");
-const qrcode = require("qrcode");
-const session = require("express-session");
-const { PrismaClient } = require("@prisma/client");
-const { SerialPort } = require("serialport");
-const WebSocket = require("ws");
-const utility = require("./src/utility");
+import chokidar from "chokidar";
+import _colors from "colors";
+import express from "express";
+import fs from "fs";
+import helmet from "helmet";
+import http from "http";
+import io from "@pm2/io";
+import { networkInterfaces } from "os";
+import path from "path";
+import qrcode from "qrcode";
+import session from "express-session";
+import { SerialPort } from "serialport";
+import { WebSocketServer } from "ws";
+import {
+    startPrismaStudio,
+    generateTimestamp,
+    fetchDefaultDisplayData,
+    fetchRankingData,
+    fetchStandaloneData,
+    fetchTimeData,
+    fetchTimes,
+    startRun,
+    endRun,
+    saveRun,
+    saveDrivers,
+    deleteTime,
+    deleteTimestamp,
+    lastVehicle,
+    reset,
+    createTimestamp,
+    fetchOperationData,
+    fetchSettingsData,
+} from "./src/db.js";
+
+import dotenv from "dotenv";
+const CONFIG = dotenv.config().parsed;
 
 //* ----------------- VARIABLES -----------------
 
-const CONFIG = dotenv.config().parsed;
-const prisma = new PrismaClient({
-    log: ["warn", "error"],
-});
-const port = CONFIG.PORT || 3000;
-const dbPort = CONFIG.PRISMA_STUDIO_PORT || 5555;
 var portOpened = null;
 var accumulatedData = "";
 
@@ -39,7 +52,7 @@ var accumulatedData = "";
 
 const app = express();
 const server = http.createServer(app);
-const wss = new WebSocket.Server({ server });
+const wss = new WebSocketServer({ server });
 
 app.set("view engine", "pug");
 app.set("views", "views");
@@ -59,7 +72,7 @@ app.use(
     })
 );
 
-app.use(express.static(path.join(__dirname, "public")));
+app.use(express.static(path.join(process.cwd(), "public")));
 app.use(express.json());
 
 app.use(
@@ -83,21 +96,12 @@ chokidar.watch(CONFIG.DATABASE_PATH).on("change", () => {
     tellClients("reload");
 });
 
-server.listen(port, () => {
+server.listen(CONFIG.PORT, () => {
     startPrismaStudio();
-    console.log(`Server is running on port ${port}`.green);
+    console.log(`Server is running on port ${CONFIG.PORT}`.green);
 
     handleSerialPort();
 });
-
-/**
- * Error handler
- * @param {Error} error - The error to handle
- */
-errorHandler = (error) => {
-    console.error("Error occurred:", error);
-    throw error;
-};
 
 function renderView(viewName) {
     return (req, res) => {
@@ -192,20 +196,6 @@ async function handleSerialPort() {
 //* ----------------- FUNCTIONS -----------------
 
 /**
- * Generate demo timestamp for the database.
- */
-async function generateTimestamp() {
-    const time = new Date();
-
-    await prisma.timeStamp.create({
-        data: {
-            timestamp: time,
-            friendly: utility.formatTimestamp(time),
-        },
-    });
-}
-
-/**
  * Parse incoming serial data and create a new timestamp in the database.
  * @param {string} data The incoming serial data.
  * @returns {Promise<void>} A promise that resolves when the data has been parsed.
@@ -243,36 +233,12 @@ async function parseSerialData(data) {
         );
 
         accumulatedData = "";
-
-        await prisma.timeStamp.create({
-            data: {
-                timestamp: date,
-                friendly: utility.formatTimestamp(date),
-            },
-        });
+        TODO;
+        await createTimestamp();
     } else {
         console.error(`No time found in data, accumulating more data...`);
         console.error(`Accumulated data: ${accumulatedData}`);
     }
-}
-
-/**
- * Run a query using Prisma and handle the result.
- * @param {Function} query The query to run.
- * @returns {Promise} The result of the query.
- * @throws {Error} If an error occurs during the query.
- */
-async function runQuery(query) {
-    console.log("Running query.");
-    return query(prisma)
-        .then((queryResult) => {
-            return queryResult;
-        })
-        .catch(async (e) => {
-            console.error(e);
-            await prisma.$disconnect();
-            process.exit(1);
-        });
 }
 
 /**
@@ -309,26 +275,6 @@ function getAllServerIps() {
     return ipAddresses;
 }
 
-/**
- * Start Prisma Studio if it is not already running.
- */
-function startPrismaStudio() {
-    const req = http.request(
-        {
-            hostname: "localhost",
-            port: dbPort,
-        },
-        (res) => {
-            console.log(`Prisma Studio already running`);
-        }
-    );
-
-    req.on("error", (e) => {
-        console.log(`Prisma Studio not running, starting...`);
-        exec(`npx prisma studio -p ${dbPort} -b none`);
-    });
-}
-
 async function fetchInfoData() {
     const ips = getAllServerIps();
 
@@ -352,361 +298,9 @@ async function fetchInfoData() {
     };
 }
 
-async function fetchDefaultDisplayData() {
-    const lastRun = await prisma.time.findFirst({
-        where: {
-            NOT: {
-                endTime: null,
-            },
-            active: true,
-        },
-        orderBy: {
-            id: "desc",
-        },
-        include: {
-            driver: true,
-            vehicle: true,
-            startTime: true,
-            endTime: true,
-        },
-    });
-
-    if (!lastRun) {
-        return null;
-    }
-
-    lastRun.time = await utility.getTime(
-        lastRun.startTime.timestamp,
-        lastRun.endTime.timestamp,
-        lastRun.penalty
-    );
-
-    const lastRuns = await prisma.time.findMany({
-        where: {
-            driverId: lastRun.driverId,
-            NOT: {
-                endTime: null,
-            },
-            active: true,
-        },
-        orderBy: {
-            id: "desc",
-        },
-        include: {
-            driver: true,
-            vehicle: true,
-            startTime: true,
-            endTime: true,
-        },
-    });
-
-    for (let i = 0; i < lastRuns.length; i++) {
-        lastRuns[i].time = await getTime(
-            lastRuns[i].startTime.timestamp,
-            lastRuns[i].endTime.timestamp,
-            lastRuns[i].penalty
-        );
-    }
-
-    return { lastRun: lastRun, allDriverRuns: lastRuns };
-}
-
-async function fetchRankingData() {
-    const times = await prisma.time.findMany({
-        where: {
-            active: true,
-            NOT: {
-                endTime: null,
-            },
-        },
-        include: {
-            startTime: true,
-            endTime: true,
-            vehicle: true,
-            driver: true,
-        },
-    });
-
-    if (times.length === 0) {
-        return null; // Return null if no times are found
-    }
-
-    let bestTimes = new Map();
-
-    for (let time of times) {
-        let bestTime = bestTimes.get(time.driverId);
-
-        const currentTime = await utility.getTime(
-            time.startTime.timestamp,
-            time.endTime.timestamp,
-            time.penalty
-        );
-
-        if (!bestTime) {
-            bestTimes.set(time.driverId, {
-                time: currentTime,
-                driver: time.driver,
-                vehicle: time.vehicle,
-                penalty: time.penalty,
-                notes: time.notes,
-            });
-        } else {
-            if (currentTime.time < bestTime.time.time) {
-                bestTimes.set(time.driverId, {
-                    time: currentTime,
-                    driver: time.driver,
-                    vehicle: time.vehicle,
-                    penalty: time.penalty,
-                    notes: time.notes,
-                });
-            }
-        }
-    }
-
-    let ranking = Array.from(bestTimes.values()).sort(
-        (a, b) => a.time.time - b.time.time
-    );
-
-    for (let i = 0; i < ranking.length; i++) {
-        ranking[i].position = i + 1;
-    }
-
-    return { bestTimes: ranking };
-}
-
-async function fetchStandaloneData() {
-    const lastTimestamps = await prisma.timeStamp.findMany({
-        where: {
-            active: true,
-        },
-        orderBy: {
-            timestamp: "desc",
-        },
-        take: 4,
-    });
-    try {
-        main = await utility.getTime(
-            lastTimestamps[1].timestamp,
-            lastTimestamps[0].timestamp
-        ).formattedDriveTime;
-    } catch {
-        if (lastTimestamps.length == 0) {
-            return;
-        } else if (lastTimestamps.length == 1) {
-            return { main: "--:--:---", sub1: "", sub2: "" };
-        }
-    }
-    try {
-        sub1 = await utility.getTime(
-            lastTimestamps[2].timestamp,
-            lastTimestamps[1].timestamp
-        ).formattedDriveTime;
-    } catch {
-        return { main, sub1: "--:--:---", sub2: "" };
-    }
-
-    try {
-        sub2 = await utility.getTime(
-            lastTimestamps[3].timestamp,
-            lastTimestamps[2].timestamp
-        ).formattedDriveTime;
-    } catch {
-        return { main, sub1, sub2: "--:--:---" };
-    }
-    return { main, sub1, sub2 };
-}
-
-async function fetchOperationData() {
-    await handleSerialPort();
-    const timestamps = await runQuery(async (prisma) => {
-        return prisma.timeStamp.findMany({
-            where: {
-                active: true,
-            },
-            orderBy: {
-                id: "asc",
-            },
-        });
-    });
-
-    let drivers = await runQuery(async (prisma) => {
-        return prisma.driver.findMany({
-            orderBy: {
-                active: "asc",
-            },
-            where: {
-                active: { not: null },
-            },
-        });
-    });
-
-    const vehicles = await runQuery(async (prisma) => {
-        return prisma.vehicle.findMany();
-    });
-
-    const timesStarted = await runQuery(async (prisma) => {
-        return prisma.time.findMany({
-            where: {
-                endTime: null,
-                active: true,
-            },
-            include: {
-                startTime: true,
-                driver: true,
-                vehicle: true,
-            },
-        });
-    });
-
-    const timesEnded = await runQuery(async (prisma) => {
-        return prisma.time.findMany({
-            where: {
-                NOT: {
-                    endTime: null,
-                },
-                notes: null,
-                active: true,
-            },
-            include: {
-                startTime: true,
-                endTime: true,
-                driver: true,
-                vehicle: true,
-            },
-        });
-    });
-
-    for (let time of timesEnded) {
-        if (time.startTime && time.endTime) {
-            const timeData = await utility.getTime(
-                time.startTime.timestamp,
-                time.endTime.timestamp
-            );
-            time.formattedDriveTime = timeData.formattedDriveTime;
-        }
-    }
-
-    return {
-        drivers: drivers,
-        portOpened: portOpened,
-        timestamps: timestamps,
-        vehicles: vehicles,
-        timesStarted: timesStarted,
-        timesEnded: timesEnded,
-    };
-}
-
 async function fetchPrismaStudioPort() {
-    startPrismaStudio();
+    startPrismaStudio(dbPort);
     return { dbPort: dbPort };
-}
-
-async function fetchTimeData() {
-    const times = await prisma.time.findMany({
-        where: {
-            notes: null,
-            active: true,
-        },
-        include: {
-            startTime: true,
-            endTime: true,
-            driver: true,
-            vehicle: true,
-        },
-    });
-    const timestamps = await prisma.timeStamp.findMany({
-        where: {
-            active: true,
-        },
-    });
-    return { times: times, timestamps: timestamps };
-}
-
-async function fetchSettingsData() {
-    const activeDrivers = await prisma.driver.findMany({
-        where: {
-            active: { not: null },
-        },
-        orderBy: {
-            active: "asc",
-        },
-    });
-
-    const inactiveDrivers = await prisma.driver.findMany({
-        where: {
-            active: null,
-        },
-        orderBy: [
-            {
-                trainingGroup: "asc",
-            },
-            {
-                firstName: "asc",
-            },
-        ],
-    });
-
-    const time = await prisma.time.findFirst({
-        where: {
-            active: true,
-        },
-    });
-
-    const timeStamp = await prisma.timeStamp.findFirst({
-        where: {
-            active: true,
-        },
-    });
-
-    const disableResetButton = time == null && timeStamp == null;
-
-    return {
-        activeDrivers: activeDrivers,
-        inactiveDrivers: inactiveDrivers,
-        disableResetButton: disableResetButton,
-        displayMode: CONFIG.DISPLAY_MODE,
-    };
-}
-
-async function fetchTimes() {
-    const times = await prisma.time.findMany({
-        where: {
-            active: true,
-            NOT: {
-                endTime: null,
-            },
-        },
-        include: {
-            driver: true,
-            vehicle: true,
-            startTime: true,
-            endTime: true,
-        },
-        orderBy: {
-            id: "desc",
-        },
-    });
-
-    for (let i = 0; i < times.length; i++) {
-        times[i].time = await await utility.getTime(
-            times[i].startTime.timestamp,
-            times[i].endTime.timestamp,
-            times[i].penalty
-        );
-    }
-
-    return { times: times };
-}
-
-async function deleteTime(run) {
-    await prisma.time.update({
-        where: {
-            id: run,
-        },
-        data: {
-            active: false,
-        },
-    });
 }
 
 //* ----------------- SERVER GET ROUTES -----------------
@@ -773,47 +367,7 @@ app.post("/start-run", async (req, res) => {
     const vehicleId = parseInt(req.body.vehicle);
 
     try {
-        await prisma.time.create({
-            data: {
-                driver: {
-                    connect: {
-                        id: driverId,
-                    },
-                },
-                vehicle: {
-                    connect: {
-                        id: vehicleId,
-                    },
-                },
-                startTime: {
-                    connect: {
-                        timestamp: timestamp,
-                    },
-                },
-            },
-        });
-        await prisma.timeStamp.update({
-            where: {
-                timestamp: timestamp,
-            },
-            data: {
-                active: false,
-            },
-        });
-
-        await prisma.driver.update({
-            where: {
-                id: driverId,
-            },
-            data: {
-                vehicle: {
-                    connect: {
-                        id: vehicleId,
-                    },
-                },
-            },
-        });
-
+        startRun(timestamp, driverId, vehicleId);
         res.status(200).send("Run started");
     } catch (e) {
         console.error(e);
@@ -824,7 +378,6 @@ app.post("/start-run", async (req, res) => {
 
 app.post("/delete-time", async (req, res) => {
     const run = parseInt(req.body.run);
-
     try {
         await deleteTime(run);
         res.status(200).send("Time deleted");
@@ -836,18 +389,8 @@ app.post("/delete-time", async (req, res) => {
 });
 
 app.post("/get-vehicle-for-driver", async (req, res) => {
-    const driverId = parseInt(req.body.driverId);
-
-    const driver = await prisma.driver.findFirst({
-        where: {
-            id: driverId,
-        },
-        include: {
-            vehicle: true,
-        },
-    });
-
-    res.status(200).send(driver);
+    let vehicle = lastVehicle(parseInt(req.body.driverId));
+    res.status(200).send(vehicle);
 });
 
 app.post("/save-run", async (req, res) => {
@@ -860,15 +403,7 @@ app.post("/save-run", async (req, res) => {
     }
 
     try {
-        await prisma.time.update({
-            where: {
-                id: run,
-            },
-            data: {
-                penalty: penalty,
-                notes: note,
-            },
-        });
+        saveRun(run, penalty, note);
         res.status(200).send("Run saved");
     } catch (e) {
         console.error(e);
@@ -879,21 +414,7 @@ app.post("/save-run", async (req, res) => {
 
 app.post("/reset-data", async (req, res) => {
     try {
-        await prisma.time.updateMany({
-            data: {
-                active: false,
-            },
-        });
-
-        await prisma.timeStamp.updateMany({
-            where: {
-                active: true,
-            },
-            data: {
-                active: false,
-            },
-        });
-
+        reset();
         res.status(200).send("Data reset");
     } catch (e) {
         console.error(e);
@@ -905,14 +426,7 @@ app.post("/reset-data", async (req, res) => {
 app.post("/delete-timestamp", async (req, res) => {
     const timestamp = new Date(req.body.timestamp);
     try {
-        await prisma.timeStamp.update({
-            where: {
-                timestamp: timestamp,
-            },
-            data: {
-                active: false,
-            },
-        });
+        deleteTimestamp(timestamp);
         res.status(200).send("Timestamp deleted");
     } catch (e) {
         console.error(e);
@@ -924,22 +438,7 @@ app.post("/delete-timestamp", async (req, res) => {
 app.post("/save-drivers", async (req, res) => {
     const drivers = req.body.drivers;
     try {
-        await prisma.driver.updateMany({
-            data: {
-                active: null,
-            },
-        });
-
-        for (let i = 0; i < drivers.length; i++) {
-            await prisma.driver.update({
-                where: {
-                    id: parseInt(drivers[i]),
-                },
-                data: {
-                    active: i,
-                },
-            });
-        }
+        saveDrivers(drivers);
         res.status(200).send("Drivers saved");
     } catch (e) {
         console.error(e);
@@ -953,26 +452,7 @@ app.post("/end-run", async (req, res) => {
     const run = parseInt(req.body.run);
 
     try {
-        await prisma.time.update({
-            where: {
-                id: run,
-            },
-            data: {
-                endTime: {
-                    connect: {
-                        timestamp: timestamp,
-                    },
-                },
-            },
-        });
-        await prisma.timeStamp.update({
-            where: {
-                timestamp: timestamp,
-            },
-            data: {
-                active: false,
-            },
-        });
+        endRun(run, timestamp);
         res.status(200).send("Run ended");
     } catch (e) {
         console.error(e);
